# docs/swagger.yaml
definitions:
...
  api.ApprovalActionRequest:
    properties:
      action:
        description: approve, reject, escalate, comment
        type: string
      comment:
        description: Optional comment
        type: string
      expected_version:
        description: For optimistic locking
        type: integer
    type: object
...
  api.ApprovalListResponse:
    properties:
      has_more:
        type: boolean
      items:
        items:
          $ref: '#/definitions/soar.ApprovalRequest'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  soar.ApprovalAction:
    properties:
      action:
        $ref: '#/definitions/soar.ApprovalActionType'
      approval_id:
        type: string
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  soar.ApprovalActionType:
    enum:
    - approve
    - reject
    - escalate
    - comment
    type: string
    x-enum-varnames:
    - ApprovalActionApprove
    - ApprovalActionReject
    - ApprovalActionEscalate
    - ApprovalActionComment
...
  soar.ApprovalMode:
    enum:
    - any
    - all
    - majority
    type: string
    x-enum-varnames:
    - ApprovalModeAny
    - ApprovalModeAll
    - ApprovalModeMajority
  soar.ApprovalRequest:
    properties:
      alert_id:
        type: string
      approval_count:
        type: integer
      context:
        additionalProperties: true
        description: Alert details, step params, etc.
        type: object
      created_at:
        type: string
      escalation_level:
        type: integer
      execution_id:
        type: string
      expires_at:
        type: string
      id:
        type: string
      min_approvers:
        type: integer
      mode:
        $ref: '#/definitions/soar.ApprovalMode'
      playbook_id:
        type: string
      playbook_name:
        type: string
      rejection_count:
        type: integer
      requested_by:
        type: string
      required_approvers:
        items:
          type: string
        type: array
      resolved_at:
        type: string
      status:
        $ref: '#/definitions/soar.ApprovalStatus'
      step_id:
        type: string
      step_name:
        type: string
      updated_at:
        type: string
      version:
        description: For optimistic locking
        type: integer
    type: object
...
  soar.ApprovalStatus:
    enum:
    - pending
    - approved
    - rejected
    - expired
    - escalated
    - cancelled
    type: string
    x-enum-varnames:
    - ApprovalStatusPending
    - ApprovalStatusApproved
    - ApprovalStatusRejected
    - ApprovalStatusExpired
    - ApprovalStatusEscalated
    - ApprovalStatusCancelled
...
info:
  contact: {}
paths:
...
  /api/v1/approvals:
    get:
      description: Retrieve approval requests with optional filtering by status, playbook,
        alert
      parameters:
      - description: Filter by status (pending, approved, rejected, expired, escalated)
        in: query
        name: status
        type: string
      - description: Filter by playbook ID
        in: query
        name: playbook_id
        type: string
      - description: Filter by alert ID
        in: query
        name: alert_id
        type: string
      - description: Filter by approver ID
        in: query
        name: approver_id
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 50, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ApprovalListResponse'
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List approval requests
      tags:
      - approvals
  /api/v1/approvals/{id}:
    delete:
      description: Cancel a pending approval request
      parameters:
      - description: Approval request ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Approval cancelled
        "400":
          description: Invalid request or approval not pending
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Approval not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel approval request
      tags:
      - approvals
    get:
      description: Retrieve a specific approval request by ID
      parameters:
      - description: Approval request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/soar.ApprovalRequest'
        "404":
          description: Approval not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get approval request
      tags:
      - approvals
    patch:
      consumes:
      - application/json
      description: Approve, reject, escalate, or comment on an approval request
      parameters:
      - description: Approval request ID
        in: path
        name: id
        required: true
        type: string
      - description: Action to perform
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ApprovalActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated approval request
          schema:
            $ref: '#/definitions/soar.ApprovalRequest'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Approval not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict - optimistic lock failed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Process approval action
      tags:
      - approvals
  /api/v1/approvals/{id}/actions:
    get:
      description: Retrieve the history of actions for an approval request
      parameters:
      - description: Approval request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/soar.ApprovalAction'
            type: array
        "404":
          description: Approval not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get approval actions
      tags:
      - approvals
  /api/v1/approvals/expire:
    post:
      description: Manually trigger expiration check for expired approval requests
      produces:
      - application/json
      responses:
        "200":
          description: Expiration result
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Expire approvals
      tags:
      - approvals
  /api/v1/approvals/pending:
    get:
      description: Retrieve pending approval requests where the current user is an
        authorized approver
      parameters:
      - description: 'Items per page (default: 50, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Offset (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/soar.ApprovalRequest'
            type: array
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my pending approvals
      tags:
      - approvals
...
swagger: "2.0"